Add RotateString
Given a string, the string in front of the requested number of characters to move to the end of the string, such as a string abcdef in front of the 2 characters a and b move to the end of the string, so the original string into a string cdefab. Please write a function to accomplish this function, the requirements for the length of ntime the string operations complexity O(n), space complexity O(1).